requirements.txt
# LangGraph and LangChain
langgraph>=1.1.0
langchain-google-genai>=4.2.0
langchain-core>=1.2.0

# Google Cloud
google-cloud-discoveryengine>=0.13.0
google-cloud-dlp>=3.19.0

# API
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0

# Utilities
python-dotenv>=1.0.1

# Testing
pytest>=8.2.0
